recommendations/train/requirements.txt
pymongo
dnspython
python-dotenv
numpy
